Fix test setup
Gulp still needs the requireHelper and ordered requires, but having two sets of setups were loading things twice and making it hard to fix.

This commit also expects jsdom 7 and requires node stable
