+++ SpecRunner.html
@@ -70,7 +70,6 @@
 
     <script>
         window.Marionette = Backbone.Marionette = {};
-        window.slice = Array.prototype.slice;
     </script>
 
     <script src="src/features.js"></script>
