+++ test/setup/node.js
@@ -2,7 +2,7 @@ var chai = require('chai');
 var sinon = require('sinon');
 var sinonChai = require('sinon-chai');
 var chaiJq = require('chai-jq');
-var requireHelper = require('../unit/setup/require-helper');
+var requireHelper = require('./require-helper');
 
 chai.use(sinonChai);
 chai.use(chaiJq);
@@ -13,19 +13,14 @@ global.sinon = sinon;
 if (!global.document || !global.window) {
   var jsdom = require('jsdom').jsdom;
 
-  global.document = jsdom('<html><head><script></script></head><body></body></html>', null, {
+  global.document = jsdom('<html><head><script></script></head><body></body></html>', {
     FetchExternalResources   : ['script'],
-    ProcessExternalResources : ['script'],
-    MutationEvents           : '2.0',
-    QuerySelector            : false
+    ProcessExternalResources : ['script']
   });
 
-  global.window = document.parentWindow;
+  global.window = document.defaultView;
   global.navigator = global.window.navigator;
 
-  global.window.Node.prototype.contains = function(node) {
-    return this.compareDocumentPosition(node) && 16;
-  };
 }
 
 require('babel-core/register');
@@ -35,6 +30,46 @@ global.$ = global.jQuery = require('jquery');
 global._ = require('underscore');
 global.Backbone = require('backbone');
 global.Backbone.$ = global.$;
-global.Metal = require('backbone-metal');
 global.Marionette = Backbone.Marionette = {};
-global.slice = Array.prototype.slice;
+
+requireHelper('features');
+Marionette.FEATURES.class = true;
+
+global.Metal = require('backbone-metal');
+
+require('backbone.babysitter');
+require('backbone.radio');
+
+requireHelper('bind-entity-events');
+requireHelper('radio-helpers');
+requireHelper('trigger-method');
+
+requireHelper('utils/extend');
+requireHelper('utils/isNodeAttached');
+requireHelper('utils/mergeOptions');
+requireHelper('utils/getOption');
+requireHelper('utils/proxyGetOption');
+requireHelper('utils/_getValue');
+requireHelper('utils/normalizeMethods');
+requireHelper('utils/normalizeUIString');
+requireHelper('utils/normalizeUIKeys');
+requireHelper('utils/normalizeUIValues');
+requireHelper('utils/actAsCollection');
+requireHelper('utils/deprecate');
+
+requireHelper('dom-refresh');
+requireHelper('metal');
+requireHelper('object');
+requireHelper('app-router');
+requireHelper('application');
+requireHelper('renderer');
+requireHelper('template-cache');
+requireHelper('abstract-view');
+requireHelper('view');
+requireHelper('collection-view');
+requireHelper('composite-view');
+requireHelper('behavior');
+requireHelper('behaviors');
+requireHelper('region');
+requireHelper('region-manager');
+requireHelper('error');
