+++ test/setup/require-helper.js
@@ -1,0 +1,3 @@
+module.exports = function(path) {
+  return require((process.env.APP_DIR_FOR_CODE_COVERAGE || '../../src/') + path);
+};
