+++ test/setup/setup.js
@@ -1,15 +1,41 @@
 module.exports = function() {
   global.expect = global.chai.expect;
 
+  var $fixtures;
+
+  function setFixtures() {
+    _.each(arguments, function(content) {
+      $fixtures.append(content);
+    });
+  }
+
+  function clearFixtures() {
+    $fixtures.empty();
+  }
+
+  function checkProperties(block, blacklist) {
+    blacklist = blacklist ? blacklist.push('cid') : 'cid';
+    var props = _.partial(_.omit, _, blacklist);
+    block.call(this, props);
+  }
+
+  before(function() {
+    $fixtures = $('<div id="fixtures">');
+    $('body').append($fixtures);
+    this.checkProperties = checkProperties;
+    this.setFixtures   = setFixtures;
+    this.clearFixtures = clearFixtures;
+  });
+
   beforeEach(function() {
-    this.sandbox = global.sinon.sandbox.create();
-    global.stub = this.sandbox.stub.bind(this.sandbox);
-    global.spy = this.sandbox.spy.bind(this.sandbox);
+    this.sinon = global.sinon.sandbox.create();
   });
 
   afterEach(function() {
-    delete global.stub;
-    delete global.spy;
-    this.sandbox.restore();
+    this.sinon.restore();
+    window.location.hash = '';
+    Backbone.history.stop();
+    Backbone.history.handlers.length = 0;
+    clearFixtures();
   });
 };
